Day 5: Not as bad as I feared.
I'm not convinced that I understand why it's OK to just compare
elements pairwise to check that they're sorted by the partial order,
but it seemed to work.

My logic was basically that it's OK to sort by a partial order, and
you can check if a sequence is sorted by comparing every consecutive
pair of elements, therefore... But that's hardly convincing.