Happy AoC. Part two took me absolutely ages.
I ended up writing a more bruteforce version which iterated ranges,
and submitted the answer to verify that it was correct. I then ran my
nice code in lockstep with it to try to find the bug. I'm not sure I
ever did, I've just ended up with the same logic, but with the
conditions reversed, and counting cases where we start on zero,
instead of end on zero.

I was worried about the first days being too easy. Definitely not.