Day 08 in Python.
I started off trying to build a k-d tree, and gave up when it became
clear that it was a lot of code which was going to need
debugging. Then I tried using the SciPy k-d tree. It turned out that
the smallest distance between points, is not the same thing as the
nearest neighbour of each point, so that didn't work. I still think
you could do something clever for part 1, but ended up bruteforcing
it, the same as apparently everyone else.